Added personality options. A character will have their current emotions and their tendencies toward certain things.
The TextMenu now accepts an array of choices instead of strings. A choice has a message, a personality weight, and a command. Since the choice has a command in it, this logic no longer needs to be done in the front-end GUI class.

Started on a ChoicePool that strips out options based on personality weight and the character's current personality, then choices 3-7 random choices from that list.